# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: clientMORE backend API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: authRegister
      tags: [auth]
      summary: Register a new operator account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Log in with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
...
  /auth/me:
    get:
      operationId: authMe
      tags: [auth]
      summary: Get current user info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
  /auth/refresh:
    post:
      operationId: authRefresh
      tags: [auth]
      summary: Refresh JWT (sliding-session renewal)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Fresh token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                required: [token]
  /auth/logout:
    post:
      operationId: authLogout
      tags: [auth]
      summary: Invalidate session (client-side)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /admin/tenants:
    get:
      operationId: adminListTenants
      tags: [admin]
      summary: List all tenants
      security:
        - adminKey: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: plan
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Tenant list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: adminCreateTenant
      tags: [admin]
      summary: Create a tenant
      security:
        - adminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                plan:
                  type: string
                limitMessages:
                  type: integer
              required: [name, email]
      responses:
        "200":
          description: Created tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /admin/tenants/{id}:
    put:
      operationId: adminUpdateTenant
      tags: [admin]
      summary: Update a tenant
      security:
        - adminKey: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    delete:
      operationId: adminDeleteTenant
      tags: [admin]
      summary: Delete a tenant
      security:
        - adminKey: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /admin/tenants/{id}/toggle:
    post:
      operationId: adminToggleTenant
      tags: [admin]
      summary: Toggle tenant active/inactive
      security:
        - adminKey: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /admin/kpis:
    get:
      operationId: adminKpis
      tags: [admin]
      summary: Global platform KPIs
      security:
        - adminKey: []
      responses:
        "200":
          description: KPIs
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string
      required: [detail]
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        companyName:
          type: string
      required: [email, password, companyName]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthSession:
      type: object
      properties:
        token:
          type: string
        userId:
          type: integer
        email:
          type: string
        name:
          type: string
          nullable: true
        role:
          type: string
        redirectTo:
          type: string
        tenantKey:
          type: string
      required: [token, userId, email, role, redirectTo, tenantKey]
    UserInfo:
      type: object
      properties:
        userId:
          type: integer
        email:
          type: string
        name:
          type: string
          nullable: true
        role:
          type: string
        tenantKey:
          type: string
      required: [userId, email, role, tenantKey]
...
    Tenant:
      type: object
      properties:
        id:
          type: integer
        tenantKey:
          type: string
        name:
          type: string
        email:
          type: string
        plan:
          type: string
        status:
          type: string
        usedMessages:
          type: integer
        limitMessages:
          type: integer
        createdDate:
          type: string
          format: date-time
      required: [id, tenantKey, name, email, plan, status]